+++ CHANGELOG.md
@@ -10,6 +10,8 @@ see wiki for more information: [wiki](https://github.com/thmarx/cms/wiki)
 
 * **MAINTENANCE** multiple dependencies updated
 * **MAINTENANCE** maven wrapper added to project
+* **FEATURE** Developer Experience [PR-440](https://github.com/CondationCMS/cms-server/pull/440)
+* **FEATURE** Aliases for content [442](https://github.com/CondationCMS/cms-server/issues/442)
 
 ### Developer experience